[storage] second step, identify and move domain objects to storage_domain
Since these types represent items on disk, let's move them closer to the
storage
